% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_browse}
\alias{hgd_browse}
\title{Opens the URL by which a httpgd device is accessible in a browser.}
\usage{
hgd_browse(endpoint = "live", which = dev.cur())
}
\arguments{
\item{endpoint}{API endpoint}

\item{which}{Which device (id)}
}
\value{
URL
}
\description{
Opens the URL by which a httpgd device is accessible in a browser.
}
\examples{
\dontrun{

hgd()
hgd_browse() # open browser
hist(rnorm(100))
#dev.off()
}
}
