% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_generate_token}
\alias{hgd_generate_token}
\title{Generate a random alphanumeric token string. 
This is mainly used internally by httpgd.}
\usage{
hgd_generate_token(len)
}
\arguments{
\item{len}{Token length}
}
\value{
Random token string
}
\description{
Generate a random alphanumeric token string. 
This is mainly used internally by httpgd.
}
\examples{
\dontrun{

hgd_generate_token(6)
}
}
