% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_info}
\alias{hgd_info}
\title{httpgd device information.}
\usage{
hgd_info(which = dev.cur())
}
\arguments{
\item{which}{Which device (ID).}
}
\value{
List of status variables with the following named items:
\verb{$id}: Server unique ID,
\verb{$version}: httpgd and library versions.
}
\description{
Access general information of a httpgd graphics device.
This function will only work after starting a device with \code{\link[=hgd]{hgd()}}.
}
\examples{
hgd()

hgd_info()

dev.off()
}
