% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_inline}
\alias{hgd_inline}
\title{Returns a plot rendered as a SVG string.
This is similar to hgd_svg(...) but the plotting code is specified inline.
This function manages (creates and destroys) the httpgd graphics device by itself.}
\usage{
hgd_inline(code, page = 0, page_width = -1, page_height = -1, file = NA, ...)
}
\arguments{
\item{code}{Plotting code.}

\item{page}{Plot page to render. If this is set to 0, the last page will be selected.}

\item{page_width}{Width of the plot. If this is set to -1, the last width will be selected.}

\item{page_height}{Height of the plot. If this is set to -1, the last height will be selected.}

\item{file}{Filepath to save SVG. (No file will be created if this is NA)}

\item{...}{Additional parameters passed to hgd(webserver=FALSE, ...)}
}
\value{
Rendered SVG string.
}
\description{
Returns a plot rendered as a SVG string.
This is similar to hgd_svg(...) but the plotting code is specified inline.
This function manages (creates and destroys) the httpgd graphics device by itself.
}
\examples{
\dontrun{

hgd_inline({
  hist(rnorm(100))
})

s <- hgd_inline({
  plot.new()
  lines(c(0.5, 1, 0.5), c(0.5, 1, 1))
})
cat(s)
}
}
