% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_renderers}
\alias{hgd_renderers}
\title{httpgd device renderers.}
\usage{
hgd_renderers()
}
\value{
List of renderers with the following named items:
\verb{$id}: Renderer ID,
\verb{$mime}: File mime type,
\verb{$ext}: File extension,
\verb{$name}: Human readable name,
\verb{$type}: Renderer type (currently either \code{plot} or \code{other}),
\verb{$bin}: Is the file a binary blob or text.
}
\description{
Get a list of available renderers.
This function will only work after starting a device with \code{\link[=hgd]{hgd()}}.
}
\examples{

hgd_renderers()

}
