% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_state}
\alias{hgd_state}
\title{Returns status information of a httpgd graphics device.}
\usage{
hgd_state(which = dev.cur())
}
\arguments{
\item{which}{Which device (id)}
}
\value{
List of status variables
}
\description{
Returns status information of a httpgd graphics device.
}
\examples{
\dontrun{

hgd()

hgd_state()

plot(1,1)

hgd_state() 

dev.off() 
}
}
