% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testgraphic.R
\name{hgd_test_pattern}
\alias{hgd_test_pattern}
\title{Plot a test pattern that can be used to evaluate and compare graphics
devices.}
\usage{
hgd_test_pattern()
}
\value{
No return value.
}
\description{
Plot a test pattern that can be used to evaluate and compare graphics
devices.
}
\examples{
hgd_test_pattern()
}
