% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_url}
\alias{hgd_url}
\title{Returns the url by which a httpgd device is accessible.}
\usage{
hgd_url(endpoint = "live", which = dev.cur(), websockets = TRUE)
}
\arguments{
\item{endpoint}{API endpoint}

\item{which}{Which device (id)}

\item{websockets}{Use websockets}
}
\value{
URL
}
\description{
Returns the url by which a httpgd device is accessible.
}
\examples{
\dontrun{

hgd()
my_url <- hgd_url()
#dev.off()
}
}
