% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_view}
\alias{hgd_view}
\title{Open httpgd URL in the IDE.}
\usage{
hgd_view()
}
\value{
\code{viewer} function return value.
}
\description{
Global option \code{viewer} needs to be set to a function that accepts the client
URL as a parameter.
}
\details{
This function will only work after starting a device with \code{\link[=hgd]{hgd()}}.
}
\examples{
if (interactive() && !is.null(getOption("viewer"))) {
  hgd()

  hgd_view()
  hist(rnorm(100))

  dev.off()
}
}
