% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{httpgdBrowse}
\alias{httpgdBrowse}
\title{Opens the url by which a httpgd device is accessible in a browser.}
\usage{
httpgdBrowse(endpoint = "live", which = dev.cur())
}
\arguments{
\item{endpoint}{API endpoint}

\item{which}{Which device (id)}
}
\value{
URL
}
\description{
Opens the url by which a httpgd device is accessible in a browser.
}
