% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{httpgdClear}
\alias{httpgdClear}
\title{Clears all plot pages.}
\usage{
httpgdClear(which = dev.cur())
}
\arguments{
\item{which}{Which device (id)}
}
\value{
Whether there were any pages to remove.
}
\description{
Clears all plot pages.
}
