% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{httpgdCloseServer}
\alias{httpgdCloseServer}
\title{Closes a graphics device if it is a httpgd server.
This achieves the same effect as dev.off().}
\usage{
httpgdCloseServer(which = dev.cur())
}
\arguments{
\item{which}{Which device (id)}
}
\description{
Closes a graphics device if it is a httpgd server.
This achieves the same effect as dev.off().
}
