% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{httpgdState}
\alias{httpgdState}
\title{Returns status information of a httpgd graphics device.}
\usage{
httpgdState(which = dev.cur())
}
\arguments{
\item{which}{Which device (id)}
}
\value{
List of status variables
}
\description{
Returns status information of a httpgd graphics device.
}
