% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hyperrefstyle}
\alias{hyperrefstyle}
\title{Adds ANSI escape codes to string if terminal supports it.}
\usage{
hyperrefstyle(str)
}
\arguments{
\item{str}{The string to style}
}
\description{
Adds ANSI escape codes to string if terminal supports it.
}
