% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_browse}
\alias{hgd_browse}
\title{Open httpgd device URL in the browser.}
\usage{
hgd_browse(endpoint = "live", which = dev.cur())
}
\arguments{
\item{endpoint}{API endpoint.}

\item{which}{Which device (ID).}
}
\value{
URL.
}
\description{
This function will only work after starting a device with \code{\link[=hgd]{hgd()}}.
}
\examples{
\dontrun{

hgd()
hgd_browse() # open browser
hist(rnorm(100))

dev.off()
}
}
