% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_browse}
\alias{hgd_browse}
\title{Open httpgd URL in the browser.}
\usage{
hgd_browse(..., which = dev.cur(), browser = getOption("browser"))
}
\arguments{
\item{...}{Parameters passed to \code{\link[=hgd_url]{hgd_url()}}.}

\item{which}{Which device (ID).}

\item{browser}{Program to be used as HTML browser.}
}
\value{
URL.
}
\description{
This function will only work after starting a device with \code{\link[=hgd]{hgd()}}.
}
\examples{
\dontrun{

hgd()
hgd_browse() # open browser
hist(rnorm(100))

dev.off()
}
}
