% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_browse}
\alias{hgd_browse}
\title{Open httpgd URL in the browser.}
\usage{
hgd_browse(..., which = dev.cur(), browser = getOption("browser"))
}
\arguments{
\item{...}{Parameters passed to \code{\link[=hgd_url]{hgd_url()}}.}

\item{which}{Which device (ID).}

\item{browser}{Program to be used as HTML browser.}
}
\value{
URL.
}
\description{
This function will only work after starting a device with \code{\link[=hgd]{hgd()}}.
}
\examples{
if (interactive()) {
  hgd() # Initialize graphics device and start server
  hgd_browse() # Or copy the displayed link in the browser

  # Plot something
  x <- seq(0, 3 * pi, by = 0.1)
  plot(x, sin(x), type = "l")

  dev.off() # alternatively: hgd_close()
}
}
