% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_clear}
\alias{hgd_clear}
\title{Clears all plot pages.}
\usage{
hgd_clear(which = dev.cur())
}
\arguments{
\item{which}{Which device (id)}
}
\value{
Whether there were any pages to remove.
}
\description{
Clears all plot pages.
}
\examples{
\dontrun{

hgd()
plot(1,1)
hist(rnorm(100))
hgd_clear()
hist(rnorm(100))
dev.off()
}
}
