% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_clear}
\alias{hgd_clear}
\title{Clear all httpgd plot pages.}
\usage{
hgd_clear(which = dev.cur())
}
\arguments{
\item{which}{Which device (ID).}
}
\value{
Whether there were any pages to remove.
}
\description{
This function will only work after starting a device with \code{\link[=hgd]{hgd()}}.
}
\examples{
\dontrun{

hgd()
plot(1,1)
hist(rnorm(100))
hgd_clear()
hist(rnorm(100))
dev.off()
}
}
