% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_close}
\alias{hgd_close}
\title{Closes a graphics device if it is a httpgd server.
This achieves the same effect as dev.off().}
\usage{
hgd_close(which = dev.cur(), all = FALSE)
}
\arguments{
\item{which}{Which device (id)}

\item{all}{If set to true, all running httpgd devices will be closed}
}
\description{
Closes a graphics device if it is a httpgd server.
This achieves the same effect as dev.off().
}
\examples{
\dontrun{

hgd()
hgd_browse() # open browser
hist(rnorm(100))
# ...
hgd_close() # Equvalent to dev.off()

hgd()
hgd()
hgd()
hgd_close(all=TRUE)
}
}
