% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_generate_token}
\alias{hgd_generate_token}
\title{Generate random alphanumeric token.}
\usage{
hgd_generate_token(len)
}
\arguments{
\item{len}{Token length (number of characters).}
}
\value{
Random token string.
}
\description{
This is mainly used internally by httpgd, but exposed for
testing purposes.
}
\examples{
hgd_generate_token(6)
}
