% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_remove}
\alias{hgd_remove}
\title{Removes a plot page.}
\usage{
hgd_remove(page = 0, which = dev.cur())
}
\arguments{
\item{page}{Plot page to remove. If this is set to 0, the last page will be selected.}

\item{which}{Which device (id)}
}
\value{
Whether the page existed (and thereby was successfully removed).
}
\description{
Removes a plot page.
}
\examples{
\dontrun{

hgd()
plot(1,1) # page 1
hist(rnorm(100)) # page 2
hgd_remove(page=1) # remove page 1
dev.off()
}
}
