% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{hgd_url}
\alias{hgd_url}
\title{httpgd server URL.}
\usage{
hgd_url(endpoint = "live", which = dev.cur(), websockets = TRUE)
}
\arguments{
\item{endpoint}{API endpoint.}

\item{which}{Which device (ID).}

\item{websockets}{Use websockets.}
}
\value{
URL.
}
\description{
This function will only work after starting a device with \code{\link[=hgd]{hgd()}}.
}
\examples{
\dontrun{

hgd()
my_url <- hgd_url()

dev.off()
}
}
