% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpgd.R
\name{httpgd}
\alias{httpgd}
\title{Initialize httpgd graphics device and start server.}
\usage{
httpgd(
  host = "127.0.0.1",
  port = 0,
  width = 720,
  height = 576,
  bg = "white",
  pointsize = 12,
  system_fonts = list(),
  user_fonts = list(),
  cors = FALSE,
  token = TRUE,
  silent = FALSE,
  websockets = TRUE
)
}
\arguments{
\item{host}{Server hostname.}

\item{port}{Server port. If this is set to 0, an open port 
will be assigned.}

\item{width}{Graphics device width (pixels).}

\item{height}{Graphics device height (pixels).}

\item{bg}{Background color.}

\item{pointsize}{Graphics device point size.}

\item{system_fonts}{Named list of font names to be aliased with
fonts installed on your system. If unspecified, the R default
families \code{sans}, \code{serif}, \code{mono} and \code{symbol}
are aliased to the family returned by
\code{\link[systemfonts]{font_info}()}.}

\item{user_fonts}{Named list of fonts to be aliased with font files
provided by the user rather than fonts properly installed on the
system. The aliases can be fonts from the fontquiver package,
strings containing a path to a font file, or a list containing
\code{name} and \code{file} elements with \code{name} indicating
the font alias in the SVG output and \code{file} the path to a
font file.}

\item{cors}{Toggles Cross-Origin Resource Sharing (CORS) header.
When set to TRUE, CORS header will be set to "*".}

\item{token}{(Optional) security token. When set, all requests
need to include a token to be allowed. (Either in a request header 
(X-HTTPGD-TOKEN) field or as a query parameter.)
This parameter can be set to TRUE to generate a random 8 character
alphanumeric token. A random token of the specified length is generated
when it is set to a number. FALSE deactivates the token.}

\item{silent}{When set to FALSE no information will be printed to the 
console after startup.}

\item{websockets}{Use websockets.}
}
\description{
Initialize httpgd graphics device and start server.
}
